audio_quality_progect/requirements.txt
soundfile
sounddevice
grpcio 
grpcio-tools